+++ src/app/(with-header)/mypage/components/ProfileNavigation.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import { useRef } from 'react';
 import { usePathname } from 'next/navigation';
 import Link from 'next/link';
 import ProfileImage from './ProfileImage';
 import useMyPageStore from '@/stores/MyPage/useMyPageStore';
-import { useUploadProfileImage } from '@/hooks/useMyPageQueries';
+import { useProfileImageUpload } from '@/hooks/useProfileImageUpload';
 import MyUserIcon from '@assets/svg/my-user';
 import MyReservationIcon from '@assets/svg/my-reservation';
 import MyActivitiesIcon from '@assets/svg/my-activities';
@@ -20,8 +19,8 @@ import MyActivitiesDashboardIcon from '@assets/svg/my-activities-dashboard';
 export default function ProfileNavigation() {
   const { user } = useMyPageStore();
   const pathname = usePathname();
-  const uploadImageMutation = useUploadProfileImage();
-  const fileInputRef = useRef<HTMLInputElement>(null);
+  const { fileInputRef, handleImageEdit, handleFileChange } =
+    useProfileImageUpload();
 
   const menuItems = [
     { href: '/mypage/profile', icon: MyUserIcon, label: '내 정보' },
@@ -45,34 +44,6 @@ export default function ProfileNavigation() {
   //   메뉴 활성화 상태 확인
   const isActive = (href: string) => {
     return pathname === href;
-  };
-
-  // 프로필 이미지 편집 버튼
-  const handleImageEdit = () => {
-    fileInputRef.current?.click();
-  };
-
-  // 파일 선택
-  const handleFileChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    const file = event.target.files?.[0];
-    if (file) {
-      // 파일 타입 검증
-      if (!file.type.startsWith('image/')) {
-        alert('이미지 파일만 업로드 가능합니다.');
-        return;
-      }
-
-      // 파일 크기 검증
-      if (file.size > 5 * 1024 * 1024) {
-        alert('파일 크기는 5MB 이하여야 합니다.');
-        return;
-      }
-
-      uploadImageMutation.mutate(file);
-    }
-
-    // 같은 파일을 다시 선택할 수 있도록 input 값 초기화
-    event.target.value = '';
   };
 
   return (
